// packages/primitives/src/react/toast/types.ts
export interface ToastClassnames {
    toast?: string;
    title?: string;
    description?: string;
    loader?: string;
    closeButton?: string;
    cancelButton?: string;
    actionButton?: string;
    success?: string;
    error?: string;
    info?: string;
    warning?: string;
    loading?: string;
    default?: string;
    content?: string;
    icon?: string;
}

export interface ToastIcons {
    success?: React.ReactNode;
    info?: React.ReactNode;
    warning?: React.ReactNode;
    error?: React.ReactNode;
    loading?: React.ReactNode;
    close?: React.ReactNode;
}
// ...
export type Position = "top-left" | "top-right" | "bottom-left" | "bottom-right" | "top-center" | "bottom-center";
export interface HeightT {
    height: number;
    toastId: number | string;
    position: Position;
}

interface ToastOptions {
    className?: string;
    closeButton?: boolean;
    descriptionClassName?: string;
    style?: React.CSSProperties;
    cancelButtonStyle?: React.CSSProperties;
    actionButtonStyle?: React.CSSProperties;
    duration?: number;
    unstyled?: boolean;
    classNames?: ToastClassnames;
}

type Offset =
    | {
        top?: string | number;
        right?: string | number;
        bottom?: string | number;
        left?: string | number;
    }
    | string
    | number;

export interface ToasterProps {
    invert?: boolean;
    theme?: "light" | "dark" | "system";
    position?: Position;
    hotkey?: string[];
    richColors?: boolean;
    expand?: boolean;
    duration?: number;
    gap?: number;
    visibleToasts?: number;
    closeButton?: boolean;
    toastOptions?: ToastOptions;
    className?: string;
    style?: React.CSSProperties;
    offset?: Offset;
    mobileOffset?: Offset;
    dir?: "rtl" | "ltr" | "auto";
    swipeDirections?: SwipeDirection[];
    /**
     * @deprecated Please use the `icons` prop instead:
     * ```jsx
     * <Toaster
     *   icons={{ loading: <LoadingIcon /> }}
     * />
     * ```
     */
    loadingIcon?: React.ReactNode;
    icons?: ToastIcons;
    containerAriaLabel?: string;
    pauseWhenPageIsHidden?: boolean;
}

export type SwipeDirection = "top" | "right" | "bottom" | "left";